requirements.txt
# === Core ===
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12

# === Database ===
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
alembic>=1.14.0

# === Task Queue ===
celery[redis]>=5.4.0
redis>=5.2.0

# === LangGraph & LangChain ===
langgraph>=0.2.0
langchain>=1.0.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0

# === Embeddings & Vector Store ===
sentence-transformers>=3.3.0
faiss-cpu>=1.9.0

# === Document Parsing ===
pymupdf4llm>=0.0.17
mammoth>=1.8.0
markdownify>=0.14.0

# === Resilience & Performance ===
tenacity>=9.0.0
slowapi>=0.1.9
sse-starlette>=2.1.0

# === Utilities ===
python-dotenv>=1.0.0
